requirements.txt
pytest
ruff
python-dotenv